fix: corrige estructura de carpetas e imports
