+++ src/core/services/photo_service.py
@@ -1,0 +1,93 @@
+import json
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy import select
+import cloudinary.uploader
+
+from src.core.models.photo import Photo
+from src.core.models.tag import Tag
+
+
+# Define the photo transformations we allow.
+ALLOWED_TRANSFORMATIONS = {
+    "resize": [{"width": 800, "height": 600, "crop": "limit"}],
+    "grayscale": [{"effect": "grayscale"}],
+    "rotate": [{"angle": 90}],
+    "thumbnail": [{"width": 150, "height": 150, "crop": "thumb"}],
+}
+
+
+async def upload_photo(file, description, tags, transformations, user_id, db: AsyncSession):
+    transform_list = []
+    if transformations:
+        try:
+            user_transforms = json.loads(transformations)
+            for t in user_transforms:
+                allowed_t = {k: v for k, v in t.items() if k in ALLOWED_TRANSFORMATIONS}
+                if allowed_t:
+                    transform_list.append(allowed_t)
+        except json.JSONDecodeError:
+            raise ValueError("Invalid transformations format")
+
+    # Upload the file to Cloudinary
+    result = cloudinary.uploader.upload(
+        file.file,
+        transformation=[
+            {"width": 800, "height": 600, "crop": "limit"}, {"quality": "auto"}
+        ]
+    )
+
+    # Create a new Photo object
+    photo = Photo(
+        user_id=user_id,
+        url=result["secure_url"],
+        public_id=result["public_id"],
+        description=description
+    )
+    # Handle tags (max 5)
+    tags_objs = []
+    if tags:
+        tag_names = [t.strip() for t in tags.split(",")][:5]
+        for name in tag_names:
+            q = await db.execute(select(Tag).where(Tag.name == name))
+            tag = q.scalar_one_or_none()
+            if not tag:
+                tag = Tag(name=name)
+                db.add(tag)
+                await db.flush()
+
+            tags_objs.append(tag)
+
+    # Save the photo to the database
+    photo.tags.extend(tags_objs)
+    db.add(photo)
+    await db.commit()
+    await db.refresh(photo)
+
+
+
+async def transform_photo(photo: Photo, transformation: str, db: AsyncSession) -> str:
+    if transformation not in ALLOWED_TRANSFORMATIONS:
+        raise ValueError(f"Transformation not allowed. Choose one of: {list(ALLOWED_TRANSFORMATIONS.keys())}")
+    try:
+        result = cloudinary.uploader.explicit(
+            photo.public_id,
+            type="upload",
+            eager=ALLOWED_TRANSFORMATIONS[transformation],
+        )
+    except Exception as e:
+        raise ValueError(f"Cloudinary transformation failed: {str(e)}")
+
+
+    transformed_url = result["eager"][0]["secure_url"]
+    return transformed_url
+
+async def delete_photo(photo: Photo, db: AsyncSession):
+    try:
+        cloudinary.uploader.destroy(photo.public_id)
+    except Exception as e:
+        raise ValueError(f"Cloudinary deletion failed: {str(e)}")
+
+    await db.delete(photo)
+    await db.commit()
+
+
